+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
-- Added support for handling relabeling rules for `instance` label.
+- Added support for handling relabeling rules for `instance` label. (#175)
 
 ### Changed